DeFi-Lending-Contract.sol
The LendingProtocol smart contract in this code implements a simple peer-to-peer lending platform on the Ethereum blockchain. The contract allows lenders to lend credit to borrowers and borrowers to request loans from lenders.

The contract maintains a list of lender addresses and their total credit balances using a mapping called lenderBalances. It also maintains a list of borrowing requests using an array called requests. Each borrowing request is represented by a Request struct that contains the borrower's address, the requested amount, and a flag indicating whether the request is still open or has been fulfilled.

The lend() function allows a lender to lend credit to a borrower by specifying the borrower's address and the amount to be lent. The function first checks that the amount is greater than zero and the borrower's address is valid. It then adds the lender's address to the list of lenders and updates their credit balance. If the borrower is not already in the list of borrowing requests, the function adds a new request for the borrower. Finally, the function updates the borrower's balance in the requests array.

The requestLoan() function allows a borrower to request a loan by specifying the amount to be borrowed. The function creates a new borrowing request and adds it to the requests array.

The getRequests() function returns the full list of borrowing requests.

The fulfillRequest() function allows a lender to fulfill a borrowing request by specifying the index of the request to be fulfilled. The function first checks that the request index is valid and the lender has sufficient credit balance. It then marks the request as fulfilled, transfers the requested amount of credit from the lender's account to the borrower's account, and updates the lender's credit balance.

The getLenders() function returns the list of lender addresses.

The getBorrowers() function returns the list of borrower addresses with open borrowing requests.

The isLender() and isBorrower() functions check whether an address is a lender or a borrower, respectively.

Overall, this updated Solidity code provides a simple yet functional peer-to-peer lending platform that can be deployed on the Ethereum blockchain.